Systematized naming scheme in relational modules
