+++ exa/relational/__init__.py
@@ -2,15 +2,15 @@
 from exa.relational.base import commit, Base
 
 from exa.relational.session import Session, cleanup_anon_sessions
-from exa.relational.program import Program
-from exa.relational.project import Project
-from exa.relational.job import Job
+#from exa.relational.program import Program
+#from exa.relational.project import Project
+#from exa.relational.job import Job
 from exa.relational.container import Container
 from exa.relational.file import File
 
-from exa.relational.constants import Constant
-from exa.relational.isotopes import Isotope
-from exa.relational.units import (Length, Mass, Time, Current, Temperature,
-                                  Amount, Luminosity, Dose, Acceleration,
-                                  Angle, Charge, Dipole, Energy, Force,
-                                  Frequency, MolarMass)
+#from exa.relational.constants import Constant
+#from exa.relational.isotopes import Isotope
+#from exa.relational.units import (Length, Mass, Time, Current, Temperature,
+#                                  Amount, Luminosity, Dose, Acceleration,
+#                                  Angle, Charge, Dipole, Energy, Force,
+#                                  Frequency, MolarMass)
