+++ exa/relational/base.py
@@ -3,19 +3,52 @@
 Base Relational Objects
 ===============================================
 '''
-from sqlalchemy import Column, Integer, DateTime, Float, String
-from sqlalchemy import ForeignKey, Table
-from sqlalchemy import event, create_engine
-from sqlalchemy.orm import sessionmaker, scoped_session, mapper, relationship
-from sqlalchemy.ext.declarative import as_declarative, declared_attr, DeclarativeMeta
 from datetime import datetime
+from traitlets import MetaHasTraits
+from sqlalchemy import Column, Integer, create_engine
+from sqlalchemy.orm import sessionmaker
+from sqlalchemy.ext.declarative import as_declarative, declared_attr, DeclarativeMeta
 from exa import Config
 from exa import _pd as pd
 from exa.relational.errors import PrimaryKeyError, NameKeyError, MultipleObjectsError, NoObjectsError
 
 
-class Meta(DeclarativeMeta):
+class Meta(MetaHasTraits, DeclarativeMeta):
     '''
+    Metaclass for relational objects that allows certain relational objects
+    to act not only as a relational table schema and table entry, via
+    `sqlalchemy`_, but also as a `Jupyter notebook`_ `widget`_.
+
+    Tip:
+        Combination of the metaclasses (like this) is required because the
+        metaclass of a derived class must be a subclass of the metaclasses of
+        all of its bases. As an example,
+
+        .. code-block:: Python
+
+            class Meta:
+                pass
+
+            class Klass(object, metaclass=Meta):
+                pass
+
+        doesn't work because the metaclass (Meta) is not a subclass of "object"'s
+        metaclass (type).
+
+        .. code-block:: Python
+
+            class Meta(type):
+                pass
+
+            class Klass(object, metaclass=Meta):
+                pass
+
+        Now that the custom metaclass (Meta) has subclassed the metaclass of
+        object, the Klass class object can be created.
+
+    .. _sqlalchemy: http://www.sqlalchemy.org/
+    .. _Jupyter notebook: http://jupyter.org/
+    .. _widget: https://ipywidgets.readthedocs.org/en/latest/
     '''
     def __len__(self):
         commit()
@@ -127,12 +160,12 @@ def create_all():
     '''
     '''
     session.flush()
-    Base.metadata.create_all(engine)
+    DeclarativeBase.metadata.create_all(engine)
 
 
 engine_name = Config.relational_engine()
 engine = create_engine(engine_name)
-Session = sessionmaker(bind=engine)
+DBSession = sessionmaker(bind=engine)
 #TODO: SEE ISSUE #41  session = Session() vs session = scoped_session(sessionmaker(bind=engine))
 # For now, non-scoped
-session = Session()
+dbsession = DBSession()
