+++ exa/relational/constant.py
